okkkk

+++ Terminal/terminal.c
@@ -79,10 +79,10 @@ EN_terminalError_t isCardExpired(ST_cardData_t* cardData, ST_terminalData_t* ter
     uint8_t* endptr;
 
     //convert string to long integer using strtol and casting its return to uint8_t
-    uint8_t expMonth = (uint8_t)strtol(cardData->cardExpirationDate, &endptr, 10);
-    uint8_t expYear = (uint8_t)strtol(&cardData->cardExpirationDate[3], &endptr, 10);
-    uint8_t transactionMonth = (uint8_t)strtol(&termData->transactionDate[3], &endptr, 10);
-    uint8_t transactionYear = (uint8_t)strtol(&termData->transactionDate[8], &endptr, 10);
+    long expMonth = strtol(cardData->cardExpirationDate, &endptr, 10);
+    long expYear = strtol(&cardData->cardExpirationDate[3], &endptr, 10);
+    long transactionMonth = strtol(&termData->transactionDate[3], &endptr, 10);
+    long transactionYear = strtol(&termData->transactionDate[8], &endptr, 10);
 
    
     //Check if the expiry year is larger than year of transaction date so it's not expired
